Store Code Temporary

+++ src/arp_spoof/parsing.h
@@ -12,7 +12,7 @@ int parse_ethernet(const u_char *frame);
 // parse arp header data
 int parse_arp(const u_char *packet, struct arp_header *ahdr);
 // parse ip header data
-int parse_ip(const u_char *packet, struct ip *iphdr);
+u_int8_t parse_ip(const u_char *packet, struct ip *iphdr);
 
 #endif // PARSING_H
